Enable Music On App Exit
