+++ Tempo/Controllers/FeedViewController.swift
@@ -89,13 +89,7 @@ class FeedViewController: PlayerTableViewController, SongSearchDelegate, FeedFol
 		
 		//Animate appropriate cell if feed song is already playing
 		if let currentPost = playerNav.currentPost, playerNav.playingPostType == .feed {
-			let rowCount = tableView.numberOfRows(inSection: 0)
-			for row in 0 ..< rowCount {
-				if let thisCell = tableView.cellForRow(at: NSIndexPath(row: row, section: 0) as IndexPath) as? FeedTableViewCell, thisCell.postView.post == currentPost {
-					thisCell.postView.updatePlayingStatus()
-					break
-				}
-			}
+				thisCell.postView.updatePlayingStatus()
 		}
 	}
